+++ R/regEquation.R
@@ -13,7 +13,11 @@
 #' regEquation(X,M,Y,moderator,secondIndirect=TRUE)
 #' covar=list(name=c("C1","C2","C3"),label=c("ese","sex","tenure"),site=list(c("M1","Y"),"Y","Y"))
 #' regEquation(X,M,Y,moderator,covar=covar)
+#' covar=list(name=c("ese","sex","tenure"),site=list(c("M","Y"),c("M","Y"),c("M","Y")))
+#' regEquation(X="estress",M="affect",Y="withdraw",covar=covar)
 regEquation=function(X="X",M=NULL,Y="Y",moderator=list(),covar=list(),secondIndirect=FALSE){
+    # moderator=list();secondIndirect=FALSE
+    # X="estress";M="affect";Y="withdraw"
     (XM=moderator$name[str_detect2(moderator$site,"a")])
     (MY=moderator$name[str_detect2(moderator$site,"b")])
     (XY=moderator$name[str_detect2(moderator$site,"c")])
@@ -43,6 +47,22 @@ regEquation=function(X="X",M=NULL,Y="Y",moderator=list(),covar=list(),secondIndi
     temp=paste0(temp,collapse="+")
     equation[[count]]=paste(Y,"~",temp)
     equation2=paste0(unlist(equation),collapse="\n")
+
     equation2=addCovarEquation(equation2,covar=covar,prefix=NULL)
     equation2
+}
+
+#' Make a list of objects of class lm
+#' @param equations equations for linear regression
+#' @param data A data.frame
+#' @return a list of objects of class lm
+#' @importFrom stats as.formula lm
+#' @export
+eq2fit=function(equations,data){
+    eq=unlist(strsplit(equations,"\n"))
+    count=length(eq)
+    fit=lapply(1:count,function(i) {
+        lm(as.formula(eq[i]),data=data)
+    })
+    fit
 }